+++ src/tooling/piral-cli-webpack5/src/webpack/bundler-run.ts
@@ -41,7 +41,7 @@ export function runWebpack(wpConfig: webpack.Configuration, logLevel: LogLevels)
   };
 
   const updateBundle = (stats: webpack.Stats) => {
-    const file  = getOutput(stats);
+    const file = getOutput(stats);
     bundle.name = basename(file);
     bundle.requireRef = stats.compilation.outputOptions?.uniqueName;
     bundle.hash = stats.hash;